Revert "Feat/register (#7)"
This reverts commit 5926fe88c154b173977470528e00a04d290be5b1.
