+++ App.js
@@ -1,8 +1,12 @@
 import { StatusBar } from 'expo-status-bar';
+import { StyleSheet } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
+import Register from './src/screens/Register';
+import Login from './src/screens/Login';
+import Dashboard from './src/screens/Dashboard';
 import { useFonts } from 'expo-font';
 import AppLoading from 'expo-app-loading';
-import AuthState from './src/context/Authentication/AuthState';
 import {
   Poppins_100Thin,
   Poppins_100Thin_Italic,
@@ -23,7 +27,11 @@ import {
   Poppins_900Black,
   Poppins_900Black_Italic,
 } from '@expo-google-fonts/poppins';
-import StackNavigation from './src/navigations/StackNavigation';
+import Splash from './src/screens/Splash';
+import { useState } from 'react';
+
+const Stack = createNativeStackNavigator();
+
 export default function App() {
   let [fontsLoaded, error] = useFonts({
     Poppins_100Thin,
@@ -45,18 +53,42 @@ export default function App() {
     Poppins_900Black,
     Poppins_900Black_Italic,
   });
-
+  let [isReady, setIsReady] = useState(false);
   if (!fontsLoaded) {
     return <AppLoading />;
   }
 
   return (
-    <AuthState>
-      <NavigationContainer>
-        <StackNavigation />
-        <StatusBar style='auto' />
-      </NavigationContainer>
-    </AuthState>
+    <NavigationContainer>
+      <Stack.Navigator initialRouteName='Splash'>
+        <Stack.Screen
+          options={{ headerShown: false }}
+          name='Splash'
+          component={Splash}
+        />
+        <Stack.Screen
+          options={{ headerShown: false }}
+          name='Sign Up'
+          component={Register}
+        />
+        <Stack.Screen
+          options={{ headerShown: false }}
+          name='Sign In'
+          component={Login}
+        />
+        <Stack.Screen
+          options={{ headerShown: false }}
+          name='Dashboard'
+          component={Dashboard}
+        />
+      </Stack.Navigator>
+      <StatusBar style='auto' />
+    </NavigationContainer>
   );
 }
 
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+  },
+});
