+++ package.json
@@ -11,24 +11,13 @@
   "dependencies": {
     "@expo-google-fonts/inter": "^0.2.2",
     "@expo-google-fonts/poppins": "^0.2.2",
-
-    "@expo/vector-icons": "^13.0.0",
-    "@react-native-async-storage/async-storage": "^1.17.10",
-
     "@react-navigation/native": "^6.0.13",
     "@react-navigation/native-stack": "^6.9.0",
     "axios": "^0.27.2",
     "expo": "~46.0.9",
-
-    "expo-app-loading": "^2.1.0",
-    "expo-av": "~12.0.4",
-
     "expo-app-loading": "~2.1.0",
     "@expo/vector-icons": "^13.0.0",
-
     "expo-font": "~10.2.0",
-    "expo-secure-store": "~11.3.0",
-    "expo-splash-screen": "^0.16.2",
     "expo-status-bar": "~1.4.0",
     "react": "18.0.0",
     "react-native": "0.69.5",
@@ -36,7 +25,6 @@
     "react-native-screens": "~3.15.0",
     "react-navigation": "^4.4.4",
     "expo-av": "~12.0.4"
-
   },
   "devDependencies": {
     "@babel/core": "^7.12.9",
