--- src/components/Buttons/index.js
@@ -1,3 +1,0 @@
-import ButtonSecondary from './ButtonSecondary';
-
-export default ButtonSecondary;
