--- src/context/Authentication/authContext.js
@@ -1,5 +1,0 @@
-import { createContext } from 'react';
-
-const AuthContext = createContext();
-
-export default AuthContext;
