+++ src/screens/Dashboard.js
@@ -1,12 +1,8 @@
 import { StyleSheet, View } from 'react-native';
-
-import DashHeader from '../components/Header/DashHeader';
+import DashHeader from '../components/DashHeader';
 import { colors } from '../utils/colors';
 import Body from '../components/Body';
 import { spacing } from '../utils/sizes';
-import { useContext } from 'react';
-import AuthContext from '../context/Authentication/authContext';
-
 
 const Dashboard = () => {
   return (
