--- src/services/ApiManager.js
@@ -1,9 +1,0 @@
-import axios from 'axios';
-
-const ApiManager = axios.create({
-  baseURL: 'https://todobe-app.herokuapp.com/',
-  responseType: 'json',
-  withCredentials: true,
-});
-
-export default ApiManager;
