--- src/services/getToken.js
@@ -1,19 +1,0 @@
-import ApiManager from './ApiManager';
-
-export default getToken = async (data) => {
-  console.log(data);
-  try {
-    const result = await ApiManager('/login', {
-      method: 'POST',
-      headers: {
-        ContentType: 'application/x-www-form-urlencoded',
-      },
-      data: JSON.stringify(
-        `grant_type=&username=${data.username}&password=${data.password}&scope=&client_id=&client_secret=`
-      ),
-    });
-    return result;
-  } catch (error) {
-    console.log(error);
-  }
-};
