--- src/services/setUser.js
@@ -1,22 +1,0 @@
-import ApiManager from './ApiManager';
-
-export default setUser = async (data) => {
-  try {
-    const result = await ApiManager('/register', {
-      method: 'POST',
-      headers: {
-        ContentType: 'application/x-www-form-urlencoded',
-      },
-      data: JSON.stringify(
-        (username = data.username),
-        (email = data.email),
-        (password = data.password)
-      ),
-    });
-    console.log(result);
-    return result;
-  } catch (error) {
-    console.log('Failed');
-    console.log(error);
-  }
-};
